shapes: added edge source/target anchors in node shapes
By default (when the edge is first added between 2 nodes the anchor point is set to the center of the shape). Anchor points are correctly translated when the shape is moved across the scene. On the other hand the algorithm which moves the anchor points upon shape interactive resize needs to be adjusted since sometime anchors goes out of the shape.