edges: changed role chain and property assertion node to use edge ids in input order
This commits change the way inputs order is memoized in property assertion and role chain nodes. Instead of keeping track of the reference of the edge instance, we'll store the id of the edge (so we would need to retrieve the instance by the id later in order to call updateEdge()).
This change has been made so it's easier to load the inputs order from a .graphol file, since we just store a list of strings. In the previous implementation we would have the need to storing the edge reference inside the node instance, but edges are initialized after nodes when the diagram scene is loaded from a .graphol file, and that would introduce much more computation than the current solution.