refactor: changed application to store actions in main window instead of scene
This commit completely change the way actions are defined. In the previous implementation the diagram scene was declaring its own actions and signal slots. While everything was working fine, when switching across different documents the application had to detach and reattach action signals so that they could activate the correct slot. Also, there was a huge waste of memory since the code was spawning copies of objects we actually don't need since they were going to provide the very same functionality. I moved the whole action/slots logic inside the main window so that the amount of memory needed is reduced at its minimum, and the code is cleaner than before.