nodes: tune up performances for AbstractNode.paint()
This commit complete changes the way pens and brushes are computed. Before we were relying on the paint method that was interrogating internally a lot of other methods and properties consuming a lot of CPU cycles. Now we defined a standalone method to be called now and then when the state of the node needs to be refereshed.