core: fixed ui crashing during multiple edge insertion process
The UI was allowing the user to release the CTRL modifier while the edge insertion was in progress, thus resulting in the palette button being unchecked and consequently the Diagram mode was being reset to Idle. This was causing many troubles because it was leaving unconnected edges in the diagram which are not admissible by Eddy.