core: items: fix label recentering upon document font change
The introduction of customizable font sizes for nodes in 9eae21d3
introduced also a new feature that allowed for labels to preserve their
original position upon a change in the label's bounding box.
This is done to preserve the text alignment with the graphical node.
However, the previous implementation was causing labels to move slightly
from their default position due to a misused call to QGraphicsTextItem::adjustSize().

This commits fixes the behaviour introduced in 9eae21d3 by making
sure that the position is preserved upon font changes for labels.

Fixes issues described in #79 for v1.2