core: exporters: owl2: remove faulty for postfix character
Previously this was needed since postfix(string, value)
returned None in case value was already a suffix of the input string
(as it is always the case when value is the empty string).
This was causing some standard prefixes to be skipped when
exporting a project in owl2.  Now that the postfix function has
been fixed we can get rid of this test.

Fixes #103