ui: session: reorder new property restriction search offsets.
Before this commit when composing an OWL*PropertyExpression
new restrictions where placed in the first free location around
the expression according to the following search order:
6 ------ 3 ------ 5
2 ---- pos() ---- 1
8 ------ 4 ------ 7
Using this scheme, when composing domain/range, new restrictions
would tipically be placed on the right-hand side (or above)
of the property expression. We found out that in practice
this is not ideal given that  it is more common start placing
restrictions on the left-hand side of the expression first
(e.g. domain), hence requiring manually moving the restrictions
around after placement.
This commit changes the lookup to the following order:
5 ------ 3 ------ 6
1 ---- pos() ---- 2
7 ------ 4 ------ 8
so that by default the first locations searched are those on
the left-hand side of the expression, and there is tipically
one less manual intervention to perform when composing
a domain/range restriction.