core: factory: fix domain/range restriction switch enablement checks
Previous to this commit domain -> range restriction switch was always
allowed, including when there was an attribute in input and a concept
expression in ISA, which of course produces an illegal owl expression,
and range -> domain was only allowed if no attribute was in input to the
restriction node which would require manually deleting and recreating the
restriction node to get back to a domain restriction.

This commit changes the domain/range switch behaviour to the following:
* domain -> range switch is always allowed, except when there is an
  attribute in input and a concept expression in ISA. This avoids the
  issue mentioned above.
* range -> domain switch is always allowed except when there is a value
  domain in ISA. This allows the range -> domain switch even when an
  attribute is in input, but only until a value domain is connected in
  ISA with the range restriction node, at which point the switch is not
  legal anymore.

The remaining syntactic restriction are handled by the other validation
rules.

Closes #136