core, plugins, ui: type hints and initial cleanup of v1.x label-handing code
Start removing old code paths that where used to manage node labels
as IRIs. They are a mess and its about time we purge them.

Given that this removal will require touching a large portion of the
code base, this is the right time to start migrating to python 3.5+
style type hints from comment-based ones and all of the touched
modules have partially been converted to them.

This commit only touches some of the more common classes that are used
in the majority of the code base, the rest will follow gradually.