ui: properties: revert label management changes from v1.x
Currently all predicate nodes default to the generic node property
dialog due to the fact that they have a different item type
(see PropertyFactory.create).  This could become an issue if it goes
unnoticed once we merge old and new predicate node types, so
the predicate node property dialog will need to be adapted
for new nodes once the types will be merged.