core, plugins, ui: delete duplicated predicate item types
We don't need to keep around two implementations for these nodes,
as this creates confusion and makes the project difficult
to maintain.
This currently leaves import of v1 project broken, and
will have to be dealt later with a proper implementation.
Tests have been left out of the change for now.