ui: session: replace project path with new path when doing saveAs
Previous to this as a result of performing the saveAs action the active
project would refer to the initial project file, which means that
any subsequent save would still modify the original project file.
This behaviour is different from what is typically found in the
majority of other editing software, and can be confusing for the user.

Starting with this commit, when performing saveAs, the new selected
path becomes the active project file path, and any subsequent modification
will apply to the new project file.

Closes #142