core: plugin: switch to platform-specific folder for plugin install/scan
Previous to this, plugin install and lookup was performed from
a .eddy folder in the user home directory, independently of
the operating system.  While this may be acceptable on unix-like
systems, it is definitely wrong on platforms like Windows, where
there is no notion of 'dot' folders.

This commit changes the default location of additional plugin
install/scan to an appropriate platform-specific location,
obtained by querying QStandardsPaths.

The new locations are:
* A subfolder inside the roaming path on Windows:
  %APPDATA%\Eddy\plugins
* An application folder in the XDG_DATA_HOME on Linux:
  ~/.local/share/eddy/plugins
* A subfolder in the user's Library/Application Support folder on macOS:
  ~/Library/Application Support/Eddy/plugins

Starting with this change, the .eddy folder is not scanned anymore,
meaning that existing plugins will need to be reinstalled
through the ui, or manually copied in the appropriate folder.