core: exporters: owl2: restore selection for declaration axiom export
Starting with v3.0 the preference for declaration axiom has been removed
so that the export for declaration axiom always defaulted to false.
My best guess for this change is that there was a misunderstanding
that this axiom won't be needed since owlapi automatically adds
the declaration for entities appearing in axioms.
The problem is that if the entity does not participate to any axiom
(e.g. a plain class node), than the corresponding declaration
never reaches the final owl ontology, and the entity is basically lost.

Fix this by restoring the declaration axiom preference, so that
even if an entity does not appear in any axiom, it's declaration
is still exported in the owl ontology (if selected that is).

Notice that even if the declaration axiom is disabled during the
export, there will still be declarations for entities appearing
in any axiom, as these declaration are, as said above, automatically
added by owlapi.

Closes #149