tests: export: fix for InverseObjectProperties in normalized owl export
Now that the InverseObjectProperties axiom is restored, the owl
export with normalization test needs to be fixed since the axioms:

SubObjectPropertyOf(test:hasAncestor ObjectInverseOf(test:isAncestorOf))
SubObjectPropertyOf(ObjectInverseOf(test:isAncestorOf) test:hasAncestor)

are replaced with:

InverseObjectProperties(test:hasAncestor test:isAncestorOf)

and the count of axioms is off by one.

If this axioms was not inteded to be part of a normalized ontology,
then the fix must be applied to the normalization procedure.