core: exporters: owl2: fix ignored export preference for HasKey axiom.
Like all other axioms the HasKey axiom has a preference for filtering
it from the exported ontology. However this preference was not
taken into account, meaning HasKey axioms would always be exported.

This commit fixes this behaviour, and also refactors the way HasKey
nodes are translated so that subsequent translations of the same
node are cached.

Closes #156