core: exporters: owl2: refactor export of items' annotations
Previous to this all project's annotations (except for axiom annotations)
where exported by iterating over all project IRIs at the end of
the export procedure, and generating an annotation assertion for each of them.
This has the drawback that there is no control for filtering
annotations of entities that are part of diagrams excluded from
the exported OWL ontology.

This commit refactors the way annotations associated to items are exported
by moving their generation at the time where an item is exported (similarly
to what happended with the meta system), so that the diagram filtering
is applied. Remaining non-item annotations (such as ontology annotations)
are still dealt with separately.

Closes #155