core: datatypes: system: allow specifying more than one extension for File members
This allows making full use of the nameFilters property for
QFileDialog, enabling it to recognize more than one extension
for a single file type.
This can be useful for example to allow the user to provide
any of the most common extensions when serializing Owl 2 files,
depending on the serialization syntax, without forcing
the '.owl' extension, or when opening files which are commonly
serialized with more than one extension (as is the case for Owl).

To preserve compatibility with the old behaviour, the File.extension
property will now return the first specified extension (if any),
while a new File.extensions property can be used to obtain a list
(possibly empty) of all the available file type extensions.