core: owl: allow all facets for custom data properties
Before this commit, attempting to build an OWL expression of the form:

DatatypeRestriction(my:dataType <facet> [<facet])

would result in an exception being thrown due to the fact that for
custom datatypes there is no known constraining facet applicable.
While it is arguable whether to allow restricting custom datatypes
with contraining facets since there is no known facet space defined
for the datatype, the thrown exception definetely represents a bug
due to the Facet enum not being able to deal with custom datatypes.

This commit fixes this behaviour, while taking a relaxed approach
on what facets can be applied to custom datatypes, essentially
allowing the use of all possible constraining facets since even if
this has no clear semantic meaning (hence no reasoning will be possible),
it is still syntactically correct.
If at a later time we decide that this kind of expression should be
forbidden by the syntactic check, it would be enough to simply return
an empty set of facets for custom datatypes.

See: #177