plugins: ontology_importer: add initial versioning of database
Each db revision is simply an incremental number, which is
stored in the sqlite pragma `user_version`. The current version
of the plugin database is stored in the spec file.

Later releases will have to include the migration infrastructure,
as a set of SQL commands to change the schema and move data
from one version to the immediate successor.

See issue #203 for reference.