core: items: nodes: literal: fix doUpdateNodeLabel logic
Previous to this change the doUpdateNodeLabel used to replace
the label whenever a new literal was set. Change this by
setting the new value to the existing label node object.
Also, correctly set the initial label value when a literal
object is passed to the constructor. This fixes a bug
where the literal nodes loaded from a project file would
have the old default 'Empty' label text, and another bug
where editing the literal value and saving without changing
it would result in a duplicated label getting generated for
the literal node.