core: exporters: graphol_iri: use explicit int cast for coordinates and dimensions
Starting with python 3.10 implicit integer cast with loss is no longer
permitted. This change affects the way invocations to overloaded Qt
methods are bound in PyQt.
Previous to this change an invocation to the overloaded
QDomElement.setAttribute() method with a numeric type as argument
would always result in the integer version being invoked, even
with a floating-point number as argument. When running under
python 3.10 or later instead, a call to the QDomElement.setAttribute()
with a floating-point number would result in the float version of
the method to be called, resulting in the XML attribute getting serialized
with the floating-point value, which would then fail to be loaded
by the loader since we assume to be always using integer sizes
and coordinates in graphol.
This commit adds an explicit integer cast in the graphol exporter
whenever such situation might occur.