core: owl: emit sgnAnnotationRemoved after having removed and assertion
Prior to this the signal indicating the removal of an annotation assertion
would be emitted BEFORE the actual removal of the assertion from the list.
This however is problematic if any other component that listens for this
change to perform changes in the presentation of the entity and needs
the current list of annotations (such as, for example, the strikethrough text
presentation of deprecated entities.)
This commit moves the signal emission AFTER the actual removal of the
assertion.