core: functions: path: replace os.path.dirname(sys.executable) with sys._MEIPASS
PyInstaller 6.0.0 introduced a change for which the onedir build content
is now relocated in a subfolder called "_internal" by default.

This breaks the resource locator that in frozen app was using the
os.path.dirname(sys.executable) to locate the built module path.

As suggested change it to use sys._MEIPASS so that the resource
and plugin locator can effectively point to the right folders.

This fixes recent macOS builds failing to locate plugins, stylesheets,
and java resources which are now packaged in the Contents/Frameworks
path in the app bundle rather than the Contents/MacOS with the
executable as before.

See: https://pyinstaller.org/en/v6.0.0/CHANGES.html