Revert "fix obdasystems/eddy#267"
This reverts commit 55d1fb302adff24588282f1d6d504508edcaee7e.

This was due to a bug in the OWLAPI version we were using (4.2)
that was causing the functional syntax export to be synctactically
invalid if there was a literal with a newline for the rdfs:label
property, and was fixed in OWLAPI 4.5.4. We are now on OWLAPI
4.5.29 so this is not needed anymore.

Better not force removal of newlines to user-entered literals as
this could result in a frustrating issue for a user that has spent
time to format its literal values only to find out that they get
messed up during OWL 2 export.

The only acceptable cleaning is to strip the literal value to remove
leading and trailing whitespace when it is inserted in the project.

See: https://github.com/owlcs/owlapi/issues/758