core: exporters: graphol_iri: export all IRIs known to the project
This includes datatypes, facets, and top/bottom IRIs since we want to
enable support for managing non-graphical IRI in the project.
Note that due to this, when selecting a subset of the diagrams you stll
get the IRIs of those, as there is no way to distinguish between those
coming for external sources and those from excluded diagrams.